requirements.txt
boto3>=1.26.0
botocore>=1.29.0
Office365-REST-Python-Client>=2.4.0
